refactor: modularize package map cache
